# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GiveawayEntryInputNetwork = typeof GiveawayEntryInputNetwork[keyof typeof GiveawayEntryInputNetwork];


export const GiveawayEntryInputNetwork = {
  MTN: 'MTN',
  Airtel: 'Airtel',
  Glo: 'Glo',
  '9mobile': '9mobile',
} as const;

export interface GiveawayEntryInput {
  /**
     * @minLength 7
     * @maxLength 24
     */
  phoneNumber: string;
  /**
     * @minLength 2
     * @maxLength 40
     */
  state: string;
  network: GiveawayEntryInputNetwork;
  consentToTerms: true;
  consentToContact: true;
}

export type GiveawayEntryStatus = typeof GiveawayEntryStatus[keyof typeof GiveawayEntryStatus];


export const GiveawayEntryStatus = {
  received: 'received',
} as const;

export interface GiveawayEntry {
  id: string;
  status: GiveawayEntryStatus;
  reference: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}