# lib/api-zod/src/generated/types/giveawayEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GiveawayEntryStatus } from './giveawayEntryStatus';

export interface GiveawayEntry {
  id: string;
  status: GiveawayEntryStatus;
  reference: string;
  createdAt: Date;
}