# lib/api-zod/src/generated/types/giveawayEntryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GiveawayEntryInputNetwork } from './giveawayEntryInputNetwork';

export interface GiveawayEntryInput {
  /**
     * @minLength 7
     * @maxLength 24
     */
  phoneNumber: string;
  /**
     * @minLength 2
     * @maxLength 40
     */
  state: string;
  network: GiveawayEntryInputNetwork;
  consentToTerms: true;
  consentToContact: true;
}